+++ RELEASENOTES.md
@@ -22,6 +22,9 @@
         `MetadataRenderer(MetadataOutput, Looper, MetadataDecoderFactory,
         boolean)` to specify whether the renderer will output metadata early or
         in sync with the player position.
+*   DASH:
+    *   Parse ClearKey license URL from manifests
+        ([#10246](https://github.com/google/ExoPlayer/issues/10246)).
 *   UI:
     *   Ensure TalkBack announces the currently active speed option in the
         playback controls menu